removed unwanted params for GET http calls, also removed helpers/common
