+++ src/helpers/apps/uninstallApp.js
@@ -5,7 +5,7 @@ import type Transport from '@ledgerhq/hw-transport'
 import { BASE_SOCKET_URL } from 'config/constants'
 import { createDeviceSocket } from 'helpers/socket'
 
-import type { LedgerScriptParams } from 'helpers/common'
+import type { LedgerScriptParams } from 'helpers/types'
 import { createCustomErrorClass } from '../errors'
 
 const ManagerDeviceLockedError = createCustomErrorClass('ManagerDeviceLocked')
@@ -34,9 +34,11 @@ export default async function uninstallApp(
 ): Promise<*> {
   const params = {
     targetId,
-    ...app,
+    perso: app.perso,
+    deleteKey: app.delete_key,
     firmware: app.delete,
     firmwareKey: app.delete_key,
+    hash: app.hash,
   }
   const url = `${BASE_SOCKET_URL}/install?${qs.stringify(params)}`
   return remapError(createDeviceSocket(transport, url).toPromise())
