+++ src/helpers/types.js
@@ -1,0 +1,14 @@
+// @flow
+export type LedgerScriptParams = {
+  firmware?: string,
+  firmware_key?: string,
+  delete?: string,
+  delete_key?: string,
+  targetId?: string | number,
+  name: string,
+  version: string,
+  icon: string,
+  app?: number,
+  hash?: string,
+  perso?: string,
+}
